pyproject.toml
[build-system]
requires = ["setuptools>=64.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "music-league"
version = "0.1.0"
description = "Music League Data Scraper & Preference Predictor"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Music League Team"},
]

dependencies = [
    "playwright>=1.48.0",
    "beautifulsoup4>=4.12.3",
    "lxml>=5.3.0",
    "python-dotenv>=1.0.1",
    "tqdm>=4.66.5",
    "tenacity>=9.0.0",
    "tabulate>=0.9.0",
    "anthropic>=0.40.0",
    "spotipy>=2.24.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "fuzzywuzzy>=0.18.0",
    "python-Levenshtein>=0.21.0",
    "nltk>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
scout = "bin.scout:main"
scraper = "bin.scraper:main"
reports = "bin.reports:main"

[tool.setuptools]
packages = ["music_league"]
package-dir = {"" = "src"}

[tool.setuptools.package-data]
music_league = ["*.json", "*.txt", "*.md", "data/*"]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true